Adds child event to events file
